[State Sync] Update metric initialization for reboots.
